Add files via upload
To understand where to buy a room to generate profitable income and where the rooms are cheap to rent, Tableau can be used to visualize and analyze data. This involves collecting data on rental and property prices, importing the data into Tableau, creating map visualizations, and analyzing the data using various chart types and statistical analysis features. By using Tableau to identify trends and patterns in the data, stakeholders can make more informed decisions about where to invest in rental properties and where to rent rooms for the cheapest price. Highlighting skills and experience with Tableau and data analysis in the context of real estate can be valuable for resumes and professional development.

Link for project file :
https://drive.google.com/drive/folders/17_9qnArVGkfXAyVA76xWd0Ji0ShIDzjV?usp=share_link